# contracts/web-api.openapi.json
{
  "info": {
    "title": "Talaria Web API",
    "version": "1",
    "description": "HTTP and SSE contract of the Talaria Web server. Generated from @maudecode/talaria-web-contracts; do not edit by hand."
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/chat/stream": {
      "get": {
        "summary": "Live agent-turn relay (SSE). Journal-backed frames carry `id: <stream_id>:<seq>`; resume with `after_event_id`, `after_seq`, or `Last-Event-ID`.",
        "tags": [
          "chat"
        ],
        "parameters": [
          {
            "name": "stream_id",
            "in": "query",
            "required": true,
            "description": "Stream returned by `/api/chat/start`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after_event_id",
            "in": "query",
            "required": false,
            "description": "Resume cursor (`<stream_id>:<seq>`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "description": "Numeric resume cursor.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "replay",
            "in": "query",
            "required": false,
            "description": "`1` replays the journal from the start.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event stream; ends after `stream_end`, `cancel`, `apperror`, or `error`.",
            "content": {
              "text/event-stream": {}
            }
          },
          "404": {
            "description": "No live stream and no journal.",
            "content": {
              "application/json": {}
            }
          },
          "503": {
            "description": "Client stream limit reached (`condition: client_stream_limit`).",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/session/stream": {
      "get": {
        "summary": "Persistent per-session channel (`initial`, `server_turn_started`, `session-updated`, `bg_task_complete`).",
        "tags": [
          "chat"
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "description": "Session to follow.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "known_count",
            "in": "query",
            "required": false,
            "description": "Last message count the tab rendered; a higher persisted count triggers `session-updated`.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event stream with 5 s keepalives.",
            "content": {
              "text/event-stream": {}
            }
          }
        }
      }
    },
    "/api/sessions/events": {
      "get": {
        "summary": "Global session-list invalidation stream (`sessions_changed`).",
        "tags": [
          "sessions"
        ],
        "parameters": [
          {
            "name": "gateway",
            "in": "query",
            "required": false,
            "description": "`1` merges the gateway feed with a `stream` discriminator and an initial `gateway_status` frame.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event stream.",
            "content": {
              "text/event-stream": {}
            }
          }
        }
      }
    },
...
    "/api/approval/stream": {
      "get": {
        "summary": "Approval prompt stream (`initial` then `approval`).",
        "tags": [
          "approval"
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "description": "Session to follow.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event stream.",
            "content": {
              "text/event-stream": {}
            }
          }
        }
      }
    },
    "/api/clarify/stream": {
      "get": {
        "summary": "Clarify prompt stream (`initial` then `clarify`).",
        "tags": [
          "approval"
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "description": "Session to follow.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event stream.",
            "content": {
              "text/event-stream": {}
            }
          }
        }
      }
    },
...
  }
}